backend/requirements-minimal.txt
# Minimal requirements for testing - core functionality only
# This installs faster and helps identify if build process works

# Core Framework (Essential)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

# Azure Services (Essential)
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-cosmos==4.5.1

# Utilities (Essential)
python-dotenv==1.0.0
pydantic-settings==2.1.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Note: ML packages (torch, mmpose, etc.) removed for faster build
# Add them back once basic deployment works



